# cmd/api/server/openapi.yaml
openapi: 3.1.0
info:
  title: Code Search API
  description: |
    A self-hosted code search and replace platform with support for batch operations across repositories.

    ## Features
    - Fast indexed search across all repositories (powered by Zoekt)
    - CLI for bulk operations (search & replace with MR creation)
    - GitHub & GitLab integration (clone, sync, create PRs/MRs)

    ## Authentication
    API requests can be authenticated using a Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: Code Search Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://codesearch.example.com
    description: Production server
...
paths:
...
  /api/v1/connections:
    get:
      tags:
        - Connections
      summary: List connections
      description: Returns all configured code host connections.
      operationId: listConnections
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"

    post:
      tags:
        - Connections
      summary: Create connection
      description: Add a new code host connection (GitHub, GitLab, etc.).
      operationId: createConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionRequest"
      responses:
        "201":
          description: Connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          description: Invalid request
...
  /api/v1/connections/{id}:
    get:
      tags:
        - Connections
      summary: Get connection
      operationId: getConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Connection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          description: Connection not found

    put:
      tags:
        - Connections
      summary: Update connection
      operationId: updateConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnectionRequest"
      responses:
        "200":
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          description: Connection not found

    delete:
      tags:
        - Connections
      summary: Delete connection
      operationId: deleteConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "204":
          description: Connection deleted
        "404":
          description: Connection not found
  /api/v1/connections/{id}/test:
    post:
      tags:
        - Connections
      summary: Test connection
      description: Validate credentials and connectivity to the code host.
      operationId: testConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestConnectionResponse"
  /api/v1/connections/{id}/sync:
    post:
      tags:
        - Connections
      summary: Sync connection repositories
      description: Fetch and sync all repositories from the code host.
      operationId: syncConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncConnectionResponse"
        "409":
          description: Sync already in progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: already_syncing
                  message:
                    type: string
                    example: Connection sync job already running
  /api/v1/connections/{id}/repos:
    get:
      tags:
        - Connections
      summary: List connection repositories
      description: List all repositories for a specific connection.
      operationId: listConnectionRepos
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
...
  /api/v1/symbols/find:
    post:
      tags:
        - Symbols
      summary: Find symbol definitions
      description: Find where a symbol (function, class, variable) is defined.
      operationId: findSymbols
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindSymbolsRequest"
      responses:
        "200":
          description: Symbol definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindSymbolsResponse"
  /api/v1/symbols/refs:
    post:
      tags:
        - Symbols
      summary: Find symbol references
      description: Find all references to a symbol across the codebase.
      operationId: findRefs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindRefsRequest"
      responses:
        "200":
          description: Symbol references
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindRefsResponse"

  # Jobs Endpoints
...
components:
...
  parameters:
...
    ConnectionId:
      name: id
      in: path
      required: true
      description: Connection ID
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Repository:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: myorg/myrepo
        clone_url:
          type: string
          format: uri
          example: https://github.com/myorg/myrepo.git
        branches:
          type: array
          items:
            type: string
          description: List of indexed branches for this repository
          example: [main, develop]
        default_branch:
          type: string
          description: The default branch of the repository (fallback if branches is empty)
          example: main
        status:
          type: string
          enum: [pending, indexing, indexed, error]
        excluded:
          type: boolean
          description: Whether the repository is excluded from indexing
          example: false
        deleted:
          type: boolean
          description: |
            Whether the repository has been soft-deleted. Deleted repositories
            remain in the database but are excluded from indexing. They can be
            restored using the restore endpoint.
          example: false
        archived:
          type: boolean
          description: Whether the repository is archived on the code host
          example: false
        last_indexed:
          type: string
          format: date-time
...
    Connection:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: GitHub Production
        type:
          type: string
          enum: [github, github_enterprise, gitlab, gitea, bitbucket]
        url:
          type: string
          format: uri
          example: https://github.com
        created_at:
          type: string
          format: date-time
    CreateConnectionRequest:
      type: object
      required:
        - name
        - type
        - url
        - token
      properties:
        name:
          type: string
          example: GitHub Production
        type:
          type: string
          enum: [github, github_enterprise, gitlab, gitea, bitbucket]
        url:
          type: string
          format: uri
        token:
          type: string
          description: API token or personal access token
          format: password
    UpdateConnectionRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [github, github_enterprise, gitlab, gitea, bitbucket]
        url:
          type: string
          format: uri
        token:
          type: string
          format: password
    TestConnectionResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        error:
          type: string
    SyncConnectionResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        added:
          type: integer
          description: Number of new repositories added
        skipped:
          type: integer
          description: Number of repositories skipped (already exist or archived)
        total:
          type: integer
          description: Total repositories processed

    # Replace Schemas
...
    FindSymbolsRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Symbol name to search for
          example: MyClass
        kind:
          type: string
          description: Symbol type (function, class, method, variable, etc.)
        language:
          type: string
          description: Filter by language
        repos:
          type: array
          items:
            type: string
          description: Filter by repositories
        limit:
          type: integer
          default: 100
          description: Maximum number of results
    Symbol:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
        repo:
          type: string
        file:
          type: string
        line:
          type: integer
        column:
          type: integer
        signature:
          type: string
        language:
          type: string
    FindSymbolsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Symbol"
    FindRefsRequest:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
          description: Symbol name to find references for
        repos:
          type: array
          items:
            type: string
          description: Filter by repositories
        language:
          type: string
          description: Filter by language
        limit:
          type: integer
          default: 100
    Reference:
      type: object
      properties:
        repo:
          type: string
        file:
          type: string
        line:
          type: integer
        column:
          type: integer
        context:
          type: string
    FindRefsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Reference"

    # Job Schemas
...